removed WIP Directus Profile hook moved OldProfile back to Profile fixed both log out methods: as teacher and as student
