+++ packages/app/src/layouts/Authed.tsx
@@ -28,7 +28,7 @@ export const AuthedLayout: React.FC<React.PropsWithChildren> = ({
     studentId === null &&
     pathname !== "/caregiver/student-loader"
   ) {
-    //return <Redirect to="/caregiver/student-loader" />;
+    return <Redirect to="/caregiver/student-loader" />;
   }
   // implied else
