+++ packages/app/src/layouts/Unauthed.tsx
@@ -3,7 +3,6 @@ import { IonContent, IonPage } from "@ionic/react";
 import { Redirect } from "react-router-dom";
 import { useInterfaceLanguage } from "@/hooks/InterfaceLanguage";
 import { useClassroom } from "@/hooks/Classroom";
-import { useOldProfile } from "@/hooks/OldProfile";
 import { useProfile } from "@/hooks/Profile";
 import { useStudent } from "@/hooks/Student";
