+++ packages/app/src/pages/CountWithMe/CountWithMeIntro.tsx
@@ -2,7 +2,7 @@ import { DialogueScreen } from "@/components/DialogueScreen";
 import { IonText } from "@ionic/react";
 import { useHistory } from "react-router-dom";
 import { useLanguage } from "@/hooks/Language";
-import { useOldProfile } from "@/hooks/OldProfile";
+import { useProfile } from "@/hooks/Profile";
 import bili from "@/assets/img/bili_in_tshirt.png";
 import audio_en from "@/assets/audio/CountAudio/instruction_en.mp3";
 import audio_es from "@/assets/audio/CountAudio/instruction_es.mp3";
@@ -27,7 +27,7 @@ const audio_raw = [
 export const CountWithMeIntro: React.FC = () => {
   const {
     profile: { isInclusive },
-  } = useOldProfile();
+  } = useProfile();
   const history = useHistory();
   const { populateText } = useLanguage();
   const audios: string[] = populateText(audio_raw).map((a: any) => a.audio);
