+++ packages/app/src/pages/LoginWithClassroomCode/LoginWithClassroomCode.tsx
@@ -6,14 +6,14 @@ import { IonButton, IonCard, IonCardContent, IonText } from "@ionic/react";
 import { UnauthedHeader } from "@/components/UnauthedHeader";
 import { useForm } from "react-hook-form";
 import { useHistory } from "react-router-dom";
-import { useOldProfile } from "@/hooks/OldProfile";
+import { useProfile } from "@/hooks/Profile";
 import { useState } from "react";
 import { z } from "zod";
 
 export const LoginWithClassroomCode: React.FC = () => {
   const [errors, setErrors] = useState<string[] | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { setClassroom } = useOldProfile();
+  const { setClassroom } = useProfile();
   const history = useHistory();
   const functions = getFunctions();
   const findClassroomByClassroomCode = httpsCallable(
