+++ packages/app/src/pages/TeacherDashboard/AddStudentsClassroom.tsx
@@ -15,7 +15,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useHistory } from "react-router";
 import { firestore } from "@/components/Firebase";
 import { doc, updateDoc } from "firebase/firestore";
-import { useOldProfile } from "@/hooks/OldProfile";
+import { useProfile } from "@/hooks/Profile";
 import { FormattedMessage } from "react-intl";
 import { AddStudents } from "@/components/AddStudents";
 
@@ -25,7 +25,7 @@ export const AddStudentsClassroom: React.FC = () => {
   const {
     user: { uid },
     profile: { isImmersive, isInclusive, settingsLanguage },
-  } = useOldProfile();
+  } = useProfile();
   const ref = doc(firestore, "users", uid);
   // TODO: we shouldn't allow this straight from the app
   const updateProfile = (key: string, value: any) => {
